Restructuring directories Changing file locations Adding directory structure to imports Renaming update.lua to setup.lua Adding symlinking to setup.lua
