+++ Programs/monitor-system/data/datasource/energy-provider.lua
@@ -1,8 +1,8 @@
 -- Import section
-parser = require("parser")
-inherits = require("inherits")
-SingleBlock = require("single-block")
-local mock = require("mock-energy-provider")
+parser = require("utils.parser")
+inherits = require("utils.inherits")
+SingleBlock = require("data.datasource.single-block")
+local mock = require("data.mock.mock-energy-provider")
 --
 
 local EnergyProvider =
