+++ Programs/monitor-system/data/mock/mock-single-block.lua
@@ -1,5 +1,5 @@
 -- Import section
-new = require("new")
+new = require("utils.new")
 --
 
 local MockSingleBlock = {
