+++ Programs/monitor-system/resources/sound/mario-one-up.lua
@@ -1,5 +1,5 @@
 -- Import section
-local playTune = require("play-tune")
+local playTune = require("sound.play-tune")
 --
 
 local tune = {
